[socket] Fix: Remove: duplicated class TmnClientStream, use TmnClientSocketStream instead.

+++ socket/source/mnWinSockets.pas
@@ -489,7 +489,7 @@ begin
   aSockAddr.sin_port := htons(LookupPort(Port));
 
   aSockAddr.sin_addr.s_addr := inet_addr(PAnsiChar(Address));
-  if ((aSockAddr.sin_addr.s_addr - INADDR_NONE) = 0) or (aSockAddr.sin_addr.s_addr = SOCKET_ERROR) then
+  if ((aSockAddr.sin_addr.s_addr - INADDR_NONE) = 0) or (aSockAddr.sin_addr.s_addr = u_long(SOCKET_ERROR)) then
   begin
     aHostEnt := gethostbyname(PAnsiChar(Address));
     if aHostEnt <> nil then
